+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -5,18 +5,18 @@ import com.techcourse.dao.UserDao;
 import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
 import com.techcourse.domain.UserHistory;
-import java.sql.Connection;
-import java.sql.SQLException;
-import javax.sql.DataSource;
+import org.springframework.transaction.TransactionManager;
 
 public class UserService {
 
     private final UserDao userDao;
     private final UserHistoryDao userHistoryDao;
+    private final TransactionManager transactionManager;
 
-    public UserService(final UserDao userDao, final UserHistoryDao userHistoryDao) {
+    public UserService(UserDao userDao, UserHistoryDao userHistoryDao) {
         this.userDao = userDao;
         this.userHistoryDao = userHistoryDao;
+        transactionManager = new TransactionManager(DataSourceConfig.getInstance());
     }
 
     public User findById(final long id) {
@@ -28,34 +28,11 @@ public class UserService {
     }
 
     public void changePassword(long id, String newPassword, String createBy) {
-        Connection connection = getConnection();
-        try {
-            connection.setAutoCommit(false);
+        transactionManager.execute(connection -> {
             User user = findById(id);
             user.changePassword(newPassword);
             userDao.update(connection, user);
             userHistoryDao.log(connection, new UserHistory(user, createBy));
-            connection.commit();
-        } catch (SQLException e) {
-            rollback(connection);
-            throw new RuntimeException(e);
-        }
-    }
-
-    private Connection getConnection() {
-        try {
-            DataSource dataSource = DataSourceConfig.getInstance();
-            return dataSource.getConnection();
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    private void rollback(Connection connection) {
-        try {
-            connection.rollback();
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
-        }
+        });
     }
 }
