+++ jdbc/src/main/java/org/springframework/transaction/TransactionExecutor.java
@@ -1,0 +1,8 @@
+package org.springframework.transaction;
+
+import java.sql.Connection;
+
+public interface TransactionExecutor {
+
+    void execute(Connection connection);
+}
