+++ jdbc/src/main/java/org/springframework/transaction/TransactionManager.java
@@ -1,0 +1,52 @@
+package org.springframework.transaction;
+
+import java.sql.Connection;
+import java.sql.SQLException;
+import javax.sql.DataSource;
+
+public class TransactionManager {
+
+    private final DataSource dataSource;
+
+    public TransactionManager(DataSource dataSource) {
+        this.dataSource = dataSource;
+    }
+
+    public void execute(TransactionExecutor transactionExecutor) {
+        Connection connection = getConnection();
+        try {
+            connection.setAutoCommit(false);
+            transactionExecutor.execute(connection);
+            connection.commit();
+        } catch (SQLException e) {
+            rollback(connection);
+            throw new RuntimeException(e);
+        } finally {
+            close(connection);
+        }
+    }
+
+    private Connection getConnection() {
+        try {
+            return dataSource.getConnection();
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    private void rollback(Connection connection) {
+        try {
+            connection.rollback();
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    private void close(Connection connection) {
+        try {
+            connection.close();
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
+        }
+    }
+}
